Do not use blkid to create partition symlinks
This removes the use of blkid to find disk partitions and their
associated partition UUIDs. Instead /proc/partitions is parsed
to enumerate disk partitions and sgdisk is used to extract their
partition UUIDs. This prevents a race condition from causing a
deadlock between ceph-osd and krbd when deploying ceph-osd pods.

Change-Id: I875f8c0565ce6efb7cd2052725af37dec3bb3a59