Change set variable in openstack helm gate job
The volume variables is changed in mariadb chart.

It affects openstack helm gating job of mariadb, so we need to
change varible to
volume.use_local_path_for_single_pod_cluster.enabled.

Enable the local host path for mariadb data so if the container
is restarted during the gate jobs it can recover successfully.

Story: 2004702
Task: 33462
Depends-on: https://review.opendev.org/661955

Change-Id: Ib55779ddcbe7033da1900f5cd21de46543392d11