Enable hugepage support in HTK resources snippet
This patch simplifies the resource snippet in helm-toolkit to allow for
specifying hugepage limits. Specifically, this patch replaces the
individual checks for specific system resources (e.g., cpu, memory) by
just copying over the entire resource component as defined in a
values.yaml or a corresponding overwrite.

This change is a prerequisite for enable hugepage handling in other
charts such as openvswitch or postgresql.

Change-Id: I786ff6c7aa5fb6b08b54d2e21878551e5e1e3818