Add zuul-jobs roles to job job that uses them
Job openstack-helm-infra uses role named "start-zuul-console" that is part
of another project named "zuul/zuul-jobs". If this job is
ever used by another project as "parent job", it would fail, because
wouldn't find the role in any of the default pathes. This patch adds the
roles from zuul/zuul-jobs project, to the job that uses these roles from
the project

Change-Id: Ib3b7e0e43008b7a4f394b49b75529bfde9780d2f