Fluentd: Add deployment conditional for probes
This adds a conditional check on the deployment type of the
Fluentd chart to determine whether to enable the current liveness
and readiness probes or not.  The current probes are designed
around using fluentd as an aggregator and do not function properly
when fluentd is deployed as a daemonset.  When run as a daemonset
and configured to tail files via the tail input plugin, fluentd
will prioritize reading the entirety of those files before
processing other input types, including opening the forward source
socket required for the current probes to function correctly. This
results in scenarios where the current probes will fail when in
fact fluentd is functioning correctly.

Daemonset focused probes to come as a follow on once a proper path
forward has been determined

Change-Id: I8a164bd47ce1950e0bd6c5043713f4cde9f85d79
Signed-off-by: Steve Wilkerson <sw5822@att.com>