MariaDB: Disable SSL chain completion on ingress controller
This PS sets `--enable-ssl-chain-completion=false` for the MariaDB
ingress controller. This is the default for current versions of
the nginx-ingress-controller, but for 0.9.0 needs to be set.

If enableSSLChainCompletion is left on, nginx will attempt to
autocomplete SSL certificate chains with missing intermediate CA
certificates, causing unnecessary network and errors in pod logs.

Change-Id: I088b33fe994281dca6997baa87a6b599c3f10c14
Closes-Bug: #1835364