Moving job_rabbit_init from static to dynamic deps
The patch moves dependency for job_rabbit_init from static to
dynamic section because of the fact that in some cases users and
credentials for OpenStack services had been created in RabbitMQ
before deployment of the OpepnStack chart for an service. Values is
going to have the below structure. cinder-rabbit-init could be moved
from static to dynamic section:

values:
  dependencies:
    dynamic:
      job_rabbit_init:
        api:
          jobs:
            - cinder-rabbit-init

Change-Id: Ib2b9858262a229390f775ad831f8c50dfb4a19da