Ceph-RGW: fix file permision issue
This is to fix the issue we are facing with file permision on the file
/var/lib/ceph/bootstrap-rgw/ceph.keyring since owner of the file
will be root.

This is happening when node with rgw reboots and rgw pods fails at
init after reboot,this is happening on sinlge node deplyoments.

issue:

ceph-rgw-5db485fbd9-dv778  0/1  Init:CrashLoopBackOff   5  6m49s

logs:
+ chown -R ceph. /run/ceph/ /var/lib/ceph/bootstrap-rgw /var/lib/ceph/radosgw
/var/lib/ceph/tmp
chown: changing ownership of
'/var/lib/ceph/bootstrap-rgw/ceph.keyring': Operation not permitted

Change-Id: Idcb648c205053b2f03357b59173e70e02f28688c