Removing deprecated option usage in gatther pod logs logic
As per PR, https://github.com/kubernetes/kubernetes/pull/60210,
in kubectl get show-all option is deprecated and no longer needed.
Presumably now that's the default behavior.

Also in current logs gathering logic, we are interested in capturing
only pod names, so removing that option is harmless.

We are seeing related failures in local CI when kubectl version is
1.15.x. So removing this option.

Change-Id: I3886c792fe28bc8b80504d8c91e9524039131b15