Allow the user to modify the external network ips
In the 900-use-it.sh script, the user can define the gateway
(OSH_BR_EX_ADDR) and the subnet of the external network
(OSH_EXT_SUBNET). However, in the script that configures the gateway
correctly in br-ex, these values are hardcoded.

This patch allows the user to modify them by a env. variable, same as
the 900-use-it.sh script does

Change-Id: I25124024ebb72ff5fea8903fb393b71a052d0f93
Signed-off-by: Manuel Buil <mbuil@suse.com>