Support for Neutron Helm test cleanup and reset for abnormal failures
Rally usually cleans up all its resources in normal executions - normal test success cases and normal test failure cases. But the generic cleanup does not work well for out of the system failures like process interruptions, pod failures, disaster cleanup etc. This is a known issue in rally-openstack. - "Current generic mechanism is nice but it doesn't work enough well in real life. And in cases of existing users, persistence context and disaster cleanups it doesn't work well."

Hence, if we shall face above such issues, it is becoming impossible to run "helm test neutron" again because of the stale data and different quota limits mentioned in the values.yaml. Hence we need to purge the stale data from the "test" project as well as reset the quota limit for such scenarios.

For the normal executions, this patch has to do nothing, but for unexpected failures, this patch will purge the stale data from test project and reset the quota as defined in values.yaml for the next run.