Merge component, developer and multinode ingress deployment scripts
For simplification of deployment and maintenance, reduce the number
of scripts that do usually the same.
The important part is that when user wants to do multinode deployment,
OSH_DEPLOY_MULTINODE variable must be set before executing the script.

For now multinone script still stays, only to invoke the common one,
but in future it would be better to remove multinode scripts completely
and let user call the common script with the right argument.

Change-Id: I4dca159827581cbe8d6ded3be85e732acbf60ed1