Fix readiness probe of neutron-ovs-agent for OVS<2.10.0
A recently introduced readiness probe for neutron-ovs-agent makes use of
an OVSDB table entry 'dpdk_initialized' which does not exist in OVS
versions preceeding v2.10.0. This patch changes the readiness probe to
exit successfully if this table entry does not exit. Thereby it does not
give any guarantees for older versions of OVS, but at least allows the
readiness probe to pass.

Change-Id: Ic77c6bdd60730c1a7c5e55fdb4afc6db938f0ddb