[ceph-osd] Separate ceph-disk based deployment scripts
This is to create a different folder for ceph-disk based deplyoments
so that it will be easy to maintain when we introduce ceph-volume.

Separate folder for both the tools gives us flexibilty to develop or
fix the issues and commit the code to respective folders without breaking
other tool-based deployments.

Change-Id: Ib0099d292a8692dc6676eb5ed624d5d1ef677cfe