Replace "make all" in setup-client.sh
setup-client.sh is the script which setups the openstack client.
This script is referenced in the installation guide and many CI
jobs. More than installing the openstack client, the script builds
all the charts by calling `make all`. This script is also called in
openstack-helm-infra project CI job[1]. The `make all` may break the
CI at some situations. For example in the following case:

We make some changes of helm-toolkit to enhancement a tool function.
In the enhancement the function parameters are changed. At this
moment, we can't build some up-layer charts, because they call the
function with old parameters.

Actually, we don't have to make all in the setup-client.sh as we
build every single chart when we install the component. We just need
to build the helm-toolkit here. So this patch is to replace the
`make all` with `make helm-toolkit`

[1] https://github.com/openstack/openstack-helm-infra/blob/master/zuul.d/jobs.yaml#L314

Story: 2007291
Task: 38822

Change-Id: If5efb1a108a36281f2459594ba6424491d909580
Signed-off-by: songgongjun <gongjun.song@intel.com>