Grafana: Change to import the dashboards in json raw format
This code change is to enable grafana to use the raw json format
dashboards . This is to avoid the conversion of dashboards from json to
yaml and back to json during which the format is encountering issue.
Also this will help in adding new dashboards and maintaining the old
ones.
All the exisiting dashboards under values_override folder are updated to
use raw json format.

Change-Id: I48a7db1514857e082cecbb3b57deff9174509601