[CEPH-OSD] Move to 'OnDelete' upgrade strategy for ceph-osd daemonsets
This is to move to onDelete upgrade strategy  for ceph-osd daemonsets
so that osd upgrade cane be performed  by failure domains as current
upgrade strategy(RollingUpdate) will randomly pick the osd pods for upgrade.

This will be more helpful when we have rack based failure domains
on the ceph clusters.

This ps will add a new job called post-apply to restart
the osd pods rack by rack

- post-apply job will make sure osds gets restart rack by rack which
  will save upgrade time.
- its less/no distruptive since we are upgrading per failure domain.

also this job will be enabled only when we have OnDelete upgrade strategy in values.

Change-Id: I2e977e75616e08fee780f714bbd267743c42c74d