[Ceph-OSD] Update getting and releasing of lock
The PS moves the relese of lock out from the if-else statement.
Otherwise we are not releasing the lock in two of three cases.
And the lock will be released only if we are going though
the last "else" statement. The PS supposed to fix that.
Also it increases the timeout value.

Change-Id: I284cff06a34ddb864484d2a63780b4fdb9075a0a