Enable Cinder image-volume cache for the 'rbd1' backend
Enables and sets some sane, default configuration, based off
documentation [0] for the 'rbd1' scoped Cinder backend. When a
volume is first created from an image, a new cached image-volume
will be created that is owned by the Block Storage Internal Tenant.
Subsequent requests to create volumes from that image will clone
the cached version instead of downloading the image contents and
copying data to the volume.

[0]: https://docs.openstack.org/cinder/stein/admin/blockstorage-image-volume-cache.html

Change-Id: I1eae74b7058f0c6c0826289cf468033eccecaa8a